+++ src/allmydata/test/test_happiness.py
@@ -2,7 +2,6 @@
 
 from twisted.trial import unittest
 from allmydata.immutable import happiness_upload
-from allmydata.util.happinessutil import augmenting_path_for, residual_network
 
 
 class HappinessUtils(unittest.TestCase):
@@ -20,7 +19,7 @@ class HappinessUtils(unittest.TestCase):
         )
         flow = [[0 for _ in graph] for _ in graph]
 
-        residual, capacity = residual_network(graph, flow)
+        residual, capacity = happiness_upload.residual_network(graph, flow)
 
         # XXX no idea if these are right; hand-verify
         self.assertEqual(residual, [[1], [2], [3], []])
@@ -28,19 +27,6 @@ class HappinessUtils(unittest.TestCase):
 
 
 class Happiness(unittest.TestCase):
-
-    def test_original_easy(self):
-        shares = {'share0', 'share1', 'share2'}
-        peers = {'peer0', 'peer1'}
-        readonly_peers = set()
-        servermap = {
-            'peer0': {'share0'},
-            'peer1': {'share2'},
-        }
-        places0 = happiness_upload.HappinessUpload(peers, readonly_peers, shares, servermap).generate_mappings()
-
-        self.assertTrue('peer0' in places0['share0'])
-        self.assertTrue('peer1' in places0['share2'])
 
     def test_placement_simple(self):
 
@@ -56,14 +42,10 @@ class Happiness(unittest.TestCase):
         }
 
         places0 = happiness_upload.share_placement(peers, readonly_peers, shares, peers_to_shares)
-        places1 = happiness_upload.HappinessUpload(peers, readonly_peers, shares).generate_mappings()
 
         if False:
             print("places0")
             for k, v in places0.items():
-                print("  {} -> {}".format(k, v))
-            print("places1")
-            for k, v in places1.items():
                 print("  {} -> {}".format(k, v))
 
         self.assertEqual(
@@ -105,7 +87,6 @@ class Happiness(unittest.TestCase):
         }
 
         places0 = happiness_upload.share_placement(peers, readonly_peers, shares, peers_to_shares)
-        places1 = happiness_upload.HappinessUpload(peers, readonly_peers, shares).generate_mappings()
 
         # share N maps to peer N
         # i.e. this says that share0 should be on peer0, share1 should
